+++ Ejercicio.h
@@ -2,7 +2,6 @@
 #define Ejercicio_h
 #include <iostream>
 #include <sstream>
-#include<fstream>
 using namespace std;
 
 class Ejercicio {
@@ -10,28 +9,30 @@ class Ejercicio {
 	private:
 		string nombreEjercicio;
 		string repeticiones;
-		double peso;
+		float peso;
 
 
 	public:
 
-		Ejercicio(string, string, double);
+		Ejercicio(string, string, float);
 		Ejercicio();
 		virtual~Ejercicio();
 
 		string getNombreEjercicio();
 		string getRepeticiones();
-		double getPeso();
+		float getPeso();
 
 		void setNombreEjercicio(string);
 		void setRepeticiones(string);
-		void setPeso(double);
+		void setPeso(float);
 
 		string toString();
-
-
+                
 		friend ostream& operator<<(ostream&, const Ejercicio&);
 		friend istream& operator>>(istream&, Ejercicio&);
+
+
+
